newrequirements.txt
amqp==5.0.6
asgiref==3.3.1
billiard==3.6.4.0
cached-property==1.5.2
celery==5.0.5
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.16
django-celery-beat==2.2.0
django-cors-headers==3.7.0
django-filter==2.4.0
django-ranged-response==0.2.0
django-redis==4.12.1
django-rest-framework-mongoengine==3.4.1
django-rest-swagger==2.2.0
django-simple-captcha==0.5.13
django-timezone-field==4.1.2
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
dnspython==1.16.0
drf-extensions==0.7.0
et-xmlfile==1.1.0
eventlet==0.30.2
greenlet==1.1.0
idna==2.10
importlib-metadata==4.6.1
itypes==1.2.0
Jinja2==3.0.1
jsonpath==0.82
kombu==5.1.0
MarkupSafe==2.0.1
mongoengine==0.22.1
mysqlclient @ file:///F:/Addpackage/mysqlclient-2.0.3-cp37-cp37m-win_amd64.whl
numpy==1.20.3
openapi-codec==1.3.2
openpyxl==3.0.7
pandas==1.2.4
Pillow==8.1.0
prompt-toolkit==3.0.19
psutil==5.8.0
PyJWT==1.7.1
pymongo==3.11.3
pytesseract==0.3.7
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2021.1
redis==3.5.3
requests==2.25.1
scapy==2.4.5
selenium==3.141.0
simplejson==3.17.3
six==1.15.0
sqlparse==0.4.1
tinyec==0.3.1
typing-extensions==3.10.0.0
ua-parser==0.10.0
uritemplate==3.0.1
urllib3==1.26.5
user-agents==2.2.0
vine==5.0.0
wcwidth==0.2.5
xlrd==1.2.0
xlwt==1.3.0
zipp==3.5.0
